requirements.txt
aiofiles==23.2.0
aiohttp==3.9.1
httpx==0.25.2
loguru==0.7.2
openai==1.5.0
pydantic==2.5.1
python-dotenv==1.0.0
tqdm==4.66.1
wordfreq==3.0.3
